refactor the concent integration tests (abstract the runner)
