+++ scripts/concent_integration_tests/tests/playbooks/no_concent.py
@@ -1,0 +1,6 @@
+from scripts.concent_integration_tests.tests.playbooks.base import NodeTestPlaybook
+
+
+class NoConcent(NodeTestPlaybook):
+    provider_node_script = 'provider/no_concent'
+    requestor_node_script = 'requestor/no_concent'
