+++ scripts/concent_integration_tests/tests/test_regular_task_run.py
@@ -1,15 +1,5 @@
 #!/usr/bin/env python
-import sys
-
-from scripts.concent_integration_tests.tests.base import NodeTestPlaybook
+from scripts.concent_integration_tests.tests.playbooks import regular_run, run
 
 
-class RegularRun(NodeTestPlaybook):
-    provider_node_script = 'provider/debug'
-    requestor_node_script = 'requestor/debug'
-
-
-playbook = RegularRun.start()
-if playbook.exit_code:
-    print("exit code", playbook.exit_code)
-sys.exit(playbook.exit_code)
+run.run_playbook(regular_run.RegularRun)
